MY_WORK/backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
pywebpush>=2.0
python-dotenv>=1.0
httpx>=0.27
openai>=1.30.0
pydantic>=2.0.0
